refactor: make ResourceDeletionDeleteProcessor understandable
 move logic to Behavior classes per resource type
 move deletion exceptions in ResourceDeletionExceptions
