+++ zeebe/engine/src/main/java/io/camunda/zeebe/engine/processing/resource/ResourceDeletionDeleteProcessor.java
@@ -7,16 +7,10 @@
  */
 package io.camunda.zeebe.engine.processing.resource;
 
-import static io.camunda.zeebe.engine.state.instance.TimerInstance.NO_ELEMENT_INSTANCE;
-import static io.camunda.zeebe.protocol.ZbColumnFamilies.PROCESS_CACHE_BY_ID_AND_VERSION;
 import static io.camunda.zeebe.util.buffer.BufferUtil.bufferAsString;
 
-import io.camunda.search.filter.DecisionInstanceFilter;
-import io.camunda.search.filter.ProcessInstanceFilter;
-import io.camunda.security.api.model.CamundaAuthentication;
 import io.camunda.zeebe.engine.metrics.ProcessDefinitionMetrics;
 import io.camunda.zeebe.engine.processing.bpmn.behavior.BpmnBehaviors;
-import io.camunda.zeebe.engine.processing.common.CatchEventBehavior;
 import io.camunda.zeebe.engine.processing.deployment.StartEventSubscriptionManager;
 import io.camunda.zeebe.engine.processing.distribution.CommandDistributionBehavior;
 import io.camunda.zeebe.engine.processing.identity.AuthenticatedAuthorizedTenants;
@@ -24,49 +18,24 @@ import io.camunda.zeebe.engine.processing.identity.AuthorizedTenants;
 import io.camunda.zeebe.engine.processing.identity.authorization.AuthorizationCheckBehavior;
 import io.camunda.zeebe.engine.processing.identity.authorization.exception.ForbiddenException;
 import io.camunda.zeebe.engine.processing.identity.authorization.request.AuthorizationRequest;
+import io.camunda.zeebe.engine.processing.resource.ResourceDeletionExceptions.ActiveProcessInstancesException;
+import io.camunda.zeebe.engine.processing.resource.ResourceDeletionExceptions.NoSuchResourceException;
 import io.camunda.zeebe.engine.processing.streamprocessor.DistributedTypedRecordProcessor;
 import io.camunda.zeebe.engine.processing.streamprocessor.writers.StateWriter;
-import io.camunda.zeebe.engine.processing.streamprocessor.writers.TypedCommandWriter;
 import io.camunda.zeebe.engine.processing.streamprocessor.writers.TypedRejectionWriter;
 import io.camunda.zeebe.engine.processing.streamprocessor.writers.TypedResponseWriter;
 import io.camunda.zeebe.engine.processing.streamprocessor.writers.Writers;
-import io.camunda.zeebe.engine.state.deployment.DeployedDrg;
-import io.camunda.zeebe.engine.state.deployment.DeployedProcess;
-import io.camunda.zeebe.engine.state.deployment.PersistedDecision;
-import io.camunda.zeebe.engine.state.deployment.PersistedForm;
-import io.camunda.zeebe.engine.state.deployment.PersistedResource;
 import io.camunda.zeebe.engine.state.distribution.DistributionQueue;
-import io.camunda.zeebe.engine.state.immutable.BannedInstanceState;
 import io.camunda.zeebe.engine.state.immutable.DecisionState;
-import io.camunda.zeebe.engine.state.immutable.ElementInstanceState;
 import io.camunda.zeebe.engine.state.immutable.FormState;
 import io.camunda.zeebe.engine.state.immutable.ProcessState;
 import io.camunda.zeebe.engine.state.immutable.ProcessingState;
 import io.camunda.zeebe.engine.state.immutable.ResourceState;
 import io.camunda.zeebe.engine.state.immutable.TenantState;
-import io.camunda.zeebe.engine.state.immutable.TimerInstanceState;
-import io.camunda.zeebe.protocol.impl.encoding.MsgPackConverter;
-import io.camunda.zeebe.protocol.impl.record.value.batchoperation.BatchOperationCreationRecord;
-import io.camunda.zeebe.protocol.impl.record.value.deployment.DecisionRecord;
-import io.camunda.zeebe.protocol.impl.record.value.deployment.DecisionRequirementsRecord;
-import io.camunda.zeebe.protocol.impl.record.value.deployment.FormRecord;
-import io.camunda.zeebe.protocol.impl.record.value.deployment.ProcessRecord;
-import io.camunda.zeebe.protocol.impl.record.value.deployment.ResourceRecord;
-import io.camunda.zeebe.protocol.impl.record.value.history.HistoryDeletionRecord;
 import io.camunda.zeebe.protocol.impl.record.value.resource.ResourceDeletionRecord;
 import io.camunda.zeebe.protocol.record.RejectionType;
-import io.camunda.zeebe.protocol.record.ValueType;
-import io.camunda.zeebe.protocol.record.intent.BatchOperationIntent;
-import io.camunda.zeebe.protocol.record.intent.DecisionIntent;
-import io.camunda.zeebe.protocol.record.intent.DecisionRequirementsIntent;
-import io.camunda.zeebe.protocol.record.intent.FormIntent;
-import io.camunda.zeebe.protocol.record.intent.HistoryDeletionIntent;
-import io.camunda.zeebe.protocol.record.intent.ProcessIntent;
 import io.camunda.zeebe.protocol.record.intent.ResourceDeletionIntent;
-import io.camunda.zeebe.protocol.record.intent.ResourceIntent;
 import io.camunda.zeebe.protocol.record.value.AuthorizationResourceType;
-import io.camunda.zeebe.protocol.record.value.BatchOperationType;
-import io.camunda.zeebe.protocol.record.value.HistoryDeletionType;
 import io.camunda.zeebe.protocol.record.value.PermissionType;
 import io.camunda.zeebe.protocol.record.value.ResourceType;
 import io.camunda.zeebe.protocol.record.value.TenantOwned;
@@ -76,37 +45,28 @@ import java.util.List;
 import java.util.Optional;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.function.Function;
-import org.agrona.DirectBuffer;
-import org.agrona.concurrent.UnsafeBuffer;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 public class ResourceDeletionDeleteProcessor
     implements DistributedTypedRecordProcessor<ResourceDeletionRecord> {
 
-  private static final Logger LOG = LoggerFactory.getLogger(ResourceDeletionDeleteProcessor.class);
   private static final List<ResourceType> SUPPORTED_HISTORY_DELETION_TYPES =
       List.of(ResourceType.PROCESS_DEFINITION, ResourceType.DECISION_REQUIREMENTS);
 
   private final StateWriter stateWriter;
-  private final TypedCommandWriter commandWriter;
   private final TypedResponseWriter responseWriter;
   private final TypedRejectionWriter rejectionWriter;
   private final KeyGenerator keyGenerator;
   private final DecisionState decisionState;
   private final CommandDistributionBehavior commandDistributionBehavior;
   private final ProcessState processState;
-  private final ElementInstanceState elementInstanceState;
-  private final TimerInstanceState timerInstanceState;
-  private final BannedInstanceState bannedInstanceState;
-  private final CatchEventBehavior catchEventBehavior;
-  private final StartEventSubscriptions startEventSubscriptions;
   private final AuthorizationCheckBehavior authCheckBehavior;
-  private final StartEventSubscriptionManager startEventSubscriptionManager;
   private final FormState formState;
   private final ResourceState resourceState;
   private final TenantState tenantState;
-  private final ProcessDefinitionMetrics processDefinitionMetrics;
+  private final ProcessDeletionBehavior processDeletionBehavior;
+  private final FormDeletionBehavior formDeletionBehavior;
+  private final DecisionRequirementsDeletionBehavior decisionRequirementsDeletionBehavior;
+  private final ResourceDeletionBehavior resourceDeletionBehavior;
 
   public ResourceDeletionDeleteProcessor(
       final Writers writers,
@@ -117,27 +77,43 @@ public class ResourceDeletionDeleteProcessor
       final AuthorizationCheckBehavior authCheckBehavior,
       final ProcessDefinitionMetrics processDefinitionMetrics) {
     stateWriter = writers.state();
-    commandWriter = writers.command();
     responseWriter = writers.response();
     rejectionWriter = writers.rejection();
     this.keyGenerator = keyGenerator;
     decisionState = processingState.getDecisionState();
     this.commandDistributionBehavior = commandDistributionBehavior;
     processState = processingState.getProcessState();
-    elementInstanceState = processingState.getElementInstanceState();
-    timerInstanceState = processingState.getTimerState();
-    bannedInstanceState = processingState.getBannedInstanceState();
-    catchEventBehavior = bpmnBehaviors.catchEventBehavior();
     this.authCheckBehavior = authCheckBehavior;
-    startEventSubscriptionManager =
-        new StartEventSubscriptionManager(processingState, keyGenerator, stateWriter);
-    startEventSubscriptions =
-        new StartEventSubscriptions(
-            bpmnBehaviors.expressionProcessor(), catchEventBehavior, startEventSubscriptionManager);
     formState = processingState.getFormState();
     resourceState = processingState.getResourceState();
     tenantState = processingState.getTenantState();
-    this.processDefinitionMetrics = processDefinitionMetrics;
+
+    final var startEventSubscriptionManager =
+        new StartEventSubscriptionManager(processingState, keyGenerator, stateWriter);
+    final var startEventSubscriptions =
+        new StartEventSubscriptions(
+            bpmnBehaviors.expressionProcessor(),
+            bpmnBehaviors.catchEventBehavior(),
+            startEventSubscriptionManager);
+
+    processDeletionBehavior =
+        new ProcessDeletionBehavior(
+            stateWriter,
+            writers.command(),
+            keyGenerator,
+            processingState,
+            bpmnBehaviors.catchEventBehavior(),
+            startEventSubscriptionManager,
+            startEventSubscriptions,
+            processDefinitionMetrics);
+
+    formDeletionBehavior = new FormDeletionBehavior(stateWriter, keyGenerator);
+
+    decisionRequirementsDeletionBehavior =
+        new DecisionRequirementsDeletionBehavior(
+            stateWriter, writers.command(), keyGenerator, processingState.getDecisionState());
+
+    resourceDeletionBehavior = new ResourceDeletionBehavior(stateWriter, keyGenerator);
   }
 
   @Override
@@ -232,7 +208,7 @@ public class ResourceDeletionDeleteProcessor
           PermissionType.DELETE_PROCESS,
           bufferAsString(process.getBpmnProcessId()),
           process.getTenantId(),
-          () -> deleteProcess(process, command, eventKey));
+          () -> processDeletionBehavior.deleteProcess(process, command, eventKey));
     }
 
     final var drgOptional =
@@ -250,7 +226,9 @@ public class ResourceDeletionDeleteProcessor
           PermissionType.DELETE_DRD,
           bufferAsString(drg.getDecisionRequirementsId()),
           drg.getTenantId(),
-          () -> deleteDecisionRequirements(drg, command, eventKey));
+          () ->
+              decisionRequirementsDeletionBehavior.deleteDecisionRequirements(
+                  drg, command, eventKey));
     }
 
     final var formOptional = formState.findFormByKey(value.getResourceKey(), tenantId);
@@ -267,7 +245,7 @@ public class ResourceDeletionDeleteProcessor
           PermissionType.DELETE_FORM,
           bufferAsString(form.getFormId()),
           form.getTenantId(),
-          () -> deleteForm(form));
+          () -> formDeletionBehavior.deleteForm(form));
     }
 
     final var resourceOptional = resourceState.findResourceByKey(value.getResourceKey(), tenantId);
@@ -284,7 +262,7 @@ public class ResourceDeletionDeleteProcessor
           PermissionType.DELETE_RESOURCE,
           bufferAsString(resource.getResourceId()),
           resource.getTenantId(),
-          () -> deleteResource(resource));
+          () -> resourceDeletionBehavior.deleteResource(resource));
     }
 
     return false;
@@ -303,109 +281,6 @@ public class ResourceDeletionDeleteProcessor
     setTenantId(command, tenantId);
     deletionAction.run();
     return true;
-  }
-
-  private void deleteDecisionRequirements(
-      final DeployedDrg drg,
-      final TypedRecord<ResourceDeletionRecord> command,
-      final long eventKey) {
-    decisionState
-        .findDecisionsByTenantAndDecisionRequirementsKey(
-            drg.getTenantId(), drg.getDecisionRequirementsKey())
-        .forEach(this::deleteDecision);
-
-    if (!command.isCommandDistributed() && command.getValue().isDeleteHistory()) {
-      deleteDecisionInstanceHistory(drg.getDecisionRequirementsKey(), eventKey, command.getValue());
-    }
-
-    final var drgRecord =
-        new DecisionRequirementsRecord()
-            .setDecisionRequirementsId(bufferAsString(drg.getDecisionRequirementsId()))
-            .setDecisionRequirementsName(bufferAsString(drg.getDecisionRequirementsName()))
-            .setDecisionRequirementsVersion(drg.getDecisionRequirementsVersion())
-            .setDecisionRequirementsKey(drg.getDecisionRequirementsKey())
-            .setResourceName(bufferAsString(drg.getResourceName()))
-            .setChecksum(drg.getChecksum())
-            .setResource(drg.getResource())
-            .setTenantId(drg.getTenantId())
-            .setDeploymentKey(drg.getDeploymentKey());
-
-    stateWriter.appendFollowUpEvent(
-        keyGenerator.nextKey(), DecisionRequirementsIntent.DELETED, drgRecord);
-  }
-
-  private void deleteDecision(final PersistedDecision persistedDecision) {
-    final var decisionRecord =
-        new DecisionRecord()
-            .setDecisionId(bufferAsString(persistedDecision.getDecisionId()))
-            .setDecisionName(bufferAsString(persistedDecision.getDecisionName()))
-            .setVersion(persistedDecision.getVersion())
-            .setVersionTag(persistedDecision.getVersionTag())
-            .setDecisionKey(persistedDecision.getDecisionKey())
-            .setDecisionRequirementsId(
-                bufferAsString(persistedDecision.getDecisionRequirementsId()))
-            .setDecisionRequirementsKey(persistedDecision.getDecisionRequirementsKey())
-            .setTenantId(persistedDecision.getTenantId())
-            .setDeploymentKey(persistedDecision.getDeploymentKey());
-
-    stateWriter.appendFollowUpEvent(keyGenerator.nextKey(), DecisionIntent.DELETED, decisionRecord);
-  }
-
-  private void deleteProcess(
-      final DeployedProcess process,
-      final TypedRecord<ResourceDeletionRecord> command,
-      final long eventKey) {
-    // We don't add the checksum or resource in this event. The checksum is not easily available
-    // and the resources are left out to prevent exceeding the maximum batch size.
-    final var processIdBuffer = process.getBpmnProcessId();
-    final var tenantId = process.getTenantId();
-    final var processRecord =
-        new ProcessRecord()
-            .setBpmnProcessId(processIdBuffer)
-            .setVersion(process.getVersion())
-            .setVersionTag(process.getVersionTag())
-            .setKey(process.getKey())
-            .setResourceName(process.getResourceName())
-            .setTenantId(tenantId)
-            .setDeploymentKey(process.getDeploymentKey());
-    stateWriter.appendFollowUpEvent(keyGenerator.nextKey(), ProcessIntent.DELETING, processRecord);
-
-    final String processId = processRecord.getBpmnProcessId();
-    final var latestVersion = processState.getLatestProcessVersion(processId, tenantId);
-
-    // If we are deleting the latest version we must unsubscribe the start events
-    if (latestVersion == process.getVersion()) {
-      unsubscribeStartEvents(process);
-
-      final var previousVersion =
-          processState.findProcessVersionBefore(processId, latestVersion, tenantId);
-      // If there is a previous version we must resubscribe to the previous version's start events.
-      if (previousVersion.isPresent()) {
-        final var previousProcess =
-            processState.getProcessByProcessIdAndVersion(
-                processIdBuffer, previousVersion.get(), tenantId);
-        if (previousProcess == null) {
-          warnPreviousProcessNotFound(
-              processIdBuffer, previousVersion.get(), tenantId, processId, latestVersion);
-        } else {
-          startEventSubscriptions.resubscribeToStartEvents(previousProcess);
-        }
-      }
-    }
-
-    final var bannedInstances = bannedInstanceState.getBannedProcessInstanceKeys();
-    final var hasRunningInstances =
-        elementInstanceState.hasActiveProcessInstances(process.getKey(), bannedInstances);
-
-    if (!hasRunningInstances) {
-      if (!command.isCommandDistributed() && command.getValue().isDeleteHistory()) {
-        deleteProcessInstanceHistory(process.getKey(), eventKey, command.getValue());
-      }
-      stateWriter.appendFollowUpEvent(keyGenerator.nextKey(), ProcessIntent.DELETED, processRecord);
-      processDefinitionMetrics.processDefinitionDeleted(process.getKey());
-    } else {
-      throw new ActiveProcessInstancesException(process.getKey());
-    }
   }
 
   private void deleteHistory(
@@ -442,136 +317,16 @@ public class ResourceDeletionDeleteProcessor
 
     switch (resourceType) {
       case PROCESS_DEFINITION ->
-          deleteProcessInstanceHistory(commandValue.getResourceKey(), eventKey, commandValue);
+          processDeletionBehavior.deleteProcessInstanceHistory(
+              commandValue.getResourceKey(), eventKey, commandValue);
       case DECISION_REQUIREMENTS ->
-          deleteDecisionInstanceHistory(commandValue.getResourceKey(), eventKey, commandValue);
+          decisionRequirementsDeletionBehavior.deleteDecisionInstanceHistory(
+              commandValue.getResourceKey(), eventKey, commandValue);
       default -> {
         // No history to delete for forms and unknown resources
         // This should not be reached as SUPPORTED_HISTORY_DELETION_TYPES filters these out
       }
     }
-  }
-
-  private void deleteProcessInstanceHistory(
-      final long processDefinitionKey,
-      final long eventKey,
-      final ResourceDeletionRecord resourceDeletionRecord) {
-    final var filter =
-        new ProcessInstanceFilter.Builder().processDefinitionKeys(processDefinitionKey).build();
-    final long batchOperationKey = keyGenerator.nextKey();
-    final var batchOperationRecord =
-        new BatchOperationCreationRecord()
-            .setBatchOperationKey(batchOperationKey)
-            .setBatchOperationType(BatchOperationType.DELETE_PROCESS_INSTANCE)
-            .setEntityFilter(new UnsafeBuffer(MsgPackConverter.convertToMsgPack(filter)))
-            .setAuthentication(
-                new UnsafeBuffer(
-                    MsgPackConverter.convertToMsgPack(CamundaAuthentication.anonymous())))
-            .setFollowUpCommand(
-                ValueType.HISTORY_DELETION,
-                HistoryDeletionIntent.DELETE,
-                new HistoryDeletionRecord()
-                    .setResourceKey(processDefinitionKey)
-                    .setResourceType(HistoryDeletionType.PROCESS_DEFINITION));
-    commandWriter.appendFollowUpCommand(
-        eventKey, BatchOperationIntent.CREATE, batchOperationRecord);
-
-    resourceDeletionRecord.setBatchOperationKey(batchOperationKey);
-    resourceDeletionRecord.setBatchOperationType(BatchOperationType.DELETE_PROCESS_INSTANCE);
-  }
-
-  private void deleteDecisionInstanceHistory(
-      final long decisionRequirementsKey,
-      final long eventKey,
-      final ResourceDeletionRecord resourceDeletionRecord) {
-    final var filter =
-        new DecisionInstanceFilter.Builder()
-            .decisionRequirementsKeys(decisionRequirementsKey)
-            .build();
-    final long batchOperationKey = keyGenerator.nextKey();
-    final var batchOperationRecord =
-        new BatchOperationCreationRecord()
-            .setBatchOperationKey(batchOperationKey)
-            .setBatchOperationType(BatchOperationType.DELETE_DECISION_INSTANCE)
-            .setEntityFilter(new UnsafeBuffer(MsgPackConverter.convertToMsgPack(filter)))
-            .setAuthentication(
-                new UnsafeBuffer(
-                    MsgPackConverter.convertToMsgPack(CamundaAuthentication.anonymous())))
-            .setFollowUpCommand(
-                ValueType.HISTORY_DELETION,
-                HistoryDeletionIntent.DELETE,
-                new HistoryDeletionRecord()
-                    .setResourceKey(decisionRequirementsKey)
-                    .setResourceType(HistoryDeletionType.DECISION_REQUIREMENTS));
-    commandWriter.appendFollowUpCommand(
-        eventKey, BatchOperationIntent.CREATE, batchOperationRecord);
-
-    resourceDeletionRecord.setBatchOperationKey(batchOperationKey);
-    resourceDeletionRecord.setBatchOperationType(BatchOperationType.DELETE_DECISION_INSTANCE);
-  }
-
-  private void warnPreviousProcessNotFound(
-      final DirectBuffer processIdBuffer,
-      final int previousVersion,
-      final String tenantId,
-      final String processId,
-      final int latestVersion) {
-    LOG.warn(
-        "Expected to find previous process: {} with version: {} and tenant: '{}' to "
-            + "resubscribe start events, but no row exists in {}. "
-            + "knownVersions: {}, current latest version: {}.",
-        bufferAsString(processIdBuffer),
-        previousVersion,
-        tenantId,
-        PROCESS_CACHE_BY_ID_AND_VERSION.name(),
-        processState.getKnownProcessVersions(processId, tenantId),
-        latestVersion);
-  }
-
-  private void unsubscribeStartEvents(final DeployedProcess deployedProcess) {
-    final var process = deployedProcess.getProcess();
-    if (process.hasTimerStartEvent()) {
-      timerInstanceState.forEachTimerForElementInstance(
-          NO_ELEMENT_INSTANCE,
-          timer -> {
-            if (timer.getProcessDefinitionKey() == deployedProcess.getKey()) {
-              catchEventBehavior.unsubscribeFromTimerEvent(timer);
-            }
-          });
-    }
-
-    startEventSubscriptionManager.closeStartEventSubscriptions(deployedProcess);
-  }
-
-  private void deleteForm(final PersistedForm persistedForm) {
-    final var form =
-        new FormRecord()
-            .setFormId(persistedForm.getFormId())
-            .setFormKey(persistedForm.getFormKey())
-            .setTenantId(persistedForm.getTenantId())
-            .setResourceName(persistedForm.getResourceName())
-            .setResource(persistedForm.getResource())
-            .setChecksum(persistedForm.getChecksum())
-            .setVersion(persistedForm.getVersion())
-            .setVersionTag(persistedForm.getVersionTag())
-            .setDeploymentKey(persistedForm.getDeploymentKey());
-
-    stateWriter.appendFollowUpEvent(keyGenerator.nextKey(), FormIntent.DELETED, form);
-  }
-
-  private void deleteResource(final PersistedResource persistedResource) {
-    final var resource =
-        new ResourceRecord()
-            .setResourceId(persistedResource.getResourceId())
-            .setResourceKey(persistedResource.getResourceKey())
-            .setTenantId(persistedResource.getTenantId())
-            .setResourceName(persistedResource.getResourceName())
-            .setResource(persistedResource.getResourceBuffer())
-            .setChecksum(persistedResource.getChecksum())
-            .setVersion(persistedResource.getVersion())
-            .setVersionTag(persistedResource.getVersionTag())
-            .setDeploymentKey(persistedResource.getDeploymentKey());
-    stateWriter.appendFollowUpEvent(keyGenerator.nextKey(), ResourceIntent.DELETED, resource);
   }
 
   private AuthorizedTenants getAuthorizedTenants(
@@ -602,19 +357,11 @@ public class ResourceDeletionDeleteProcessor
     return false;
   }
 
-  /**
-   * Tries to delete the resource, iff it is assigned to the default tenant. If the resource was
-   * deleted, it returns true, otherwise false.
-   */
   private boolean tryToDeleteResourceAssignedToDefaultTenant(
       final Function<String, Boolean> resourceDeletionCallback) {
     return resourceDeletionCallback.apply(TenantOwned.DEFAULT_TENANT_IDENTIFIER);
   }
 
-  /**
-   * Loops over the existing tenants to find the resource to delete. If found and deleted, it
-   * returns true, otherwise false.
-   */
   private boolean forEachTenantUntilResourceDeleted(
       final Function<String, Boolean> resourceDeletionCallback) {
     final var resourceDeleted = new AtomicBoolean(false);
@@ -647,24 +394,6 @@ public class ResourceDeletionDeleteProcessor
             .build();
     if (authCheckBehavior.isAuthorizedOrInternalCommand(authRequest).isLeft()) {
       throw new ForbiddenException(authRequest);
-    }
-  }
-
-  private static final class NoSuchResourceException extends IllegalStateException {
-    private static final String ERROR_MESSAGE_RESOURCE_NOT_FOUND =
-        "Expected to delete resource but no resource found with key `%d`";
-
-    private NoSuchResourceException(final long resourceKey) {
-      super(String.format(ERROR_MESSAGE_RESOURCE_NOT_FOUND, resourceKey));
-    }
-  }
-
-  private static final class ActiveProcessInstancesException extends IllegalStateException {
-    private static final String ERROR_MESSAGE_RUNNING_INSTANCES =
-        "Expected to delete resource with key `%d` but there are still running instances";
-
-    private ActiveProcessInstancesException(final long processDefinitionKey) {
-      super(String.format(ERROR_MESSAGE_RUNNING_INSTANCES, processDefinitionKey));
     }
   }
 }
